requirements.txt
discord.py
requests
aiohttp
pandas
flask